+++ MA_SNN/DVSGestures/DVS_gesture_data_process/events_timeslices.py
@@ -34,9 +34,9 @@ def cast_evs(evs):
 def get_subsampled_coordinates(evs, ds_h, ds_w):
     x_coords = evs[:, 1] // ds_w
     y_coords = evs[:, 2] // ds_h
-    if x_coords.dtype != np.int:
+    if x_coords.dtype != int:
         x_coords = x_coords.astype(int)
-    if y_coords.dtype != np.int:
+    if y_coords.dtype != int:
         y_coords = y_coords.astype(int)
     return x_coords, y_coords
 
@@ -114,7 +114,7 @@ def frame_evs(times, addrs, deltat=1000, duration=500, size=[240], downsample=[1
         idx_end += find_first(times[idx_end:], t)
         if idx_end > idx_start:
             ee = addrs[idx_start:idx_end]
-            ev = [(ee[:, i] // d).astype(np.int) for i, d in enumerate(downsample)]
+            ev = [(ee[:, i] // d).astype(int) for i, d in enumerate(downsample)]
             np.add.at(chunks, tuple([i] + ev), 1)
         idx_start = idx_end
     return chunks
@@ -130,7 +130,7 @@ def chunk_evs_pol_dvs(times, addrs, deltat=1000, chunk_size=500, size=[2, 304, 2
         idx_end += find_first(times[idx_end:], t)
         if idx_end > idx_start:
             ee = addrs[idx_start:idx_end]
-            pol, x, y = ee[:, 2], (ee[:, 0] // ds_w).astype(np.int), (ee[:, 1] // ds_h).astype(np.int)
+            pol, x, y = ee[:, 2], (ee[:, 0] // ds_w).astype(int), (ee[:, 1] // ds_h).astype(int)
             np.add.at(chunks, (i, pol, x, y), 1)
         idx_start = idx_end
     return chunks
@@ -149,7 +149,7 @@ def my_chunk_evs_pol_dvs(data, dt=1000, T=500, size=[2, 304, 240], ds=[4, 4]):
         idx_end += find_first(data[idx_end:, 0], t + dt)
         if idx_end > idx_start:
             ee = data[idx_start:idx_end, 1:]
-            pol, x, y = ee[:, 0], (ee[:, 1] // ds[0]).astype(np.int), (ee[:, 2] // ds[1]).astype(np.int)
+            pol, x, y = ee[:, 0], (ee[:, 1] // ds[0]).astype(int), (ee[:, 2] // ds[1]).astype(int)
             np.add.at(chunks, (i, pol, x, y), 1)
         idx_start = idx_end
     return chunks
